break ErrorPage out and implement tests re-enable tests and coverage in builds
